+++ app/Http/Controllers/HomeController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
-// use Illuminate\Http\Request;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Highlight;
+use Illuminate\Pagination\LengthAwarePaginator;
+
 // use App\Models\Verse;
 // use Illuminate\Support\Facades\DB;
 
@@ -25,12 +27,20 @@ class HomeController extends Controller
      *
      * @return \Illuminate\Contracts\Support\Renderable
      */
-    public function index()
+    public function index(Request $request)
     {
-        $user = Auth::user();
-        $_highlights = $this->getUserHighlights($user->id);
-        return view('home', ['user'=>$user, 'highlights'=>$_highlights]);
 
+        $page = $request->query('page');
+        $user = $request->user();
+        $_highlights = $user->highlights;
+        $_highlights = $_highlights->sortBy('verse_id');
+        $_highlights = $_highlights->values();
+        // $verses = new LengthAwarePaginator($results->forPage(1, 25), count($results), 25, 1);
+        $current = $page == null ? $_highlights->forPage(1, 25) : $_highlights->forPage($page, 25);
+        // $highlights = new LengthAwarePaginator($_highlights->forPage(1, 25), count($_highlights), 25);
+        $highlights = new LengthAwarePaginator($current, count($_highlights), 25);
+        $highlights->withPath('/home');
+        return view('home', ['user'=>$user, 'highlights'=>$highlights]);
     }
 
     private function getUserHighlights($_userId) {
@@ -43,4 +53,8 @@ class HomeController extends Controller
             ->get();
             return $_highlights;
     }
+
+    // private function getUserHighlights ($user) {
+
+    // }
 }